backend/python/requirements.txt
# Core dependencies with flexible version ranges
huggingface_hub>=0.16.4,<1.0
transformers>=4.33.0,<4.36.0
sentence-transformers>=2.2.0,<2.4.0
tokenizers>=0.13.0,<0.15.0

# JSON handling
simplejson==3.19.1

# FAISS for vector database
faiss-cpu==1.7.4


# PyTorch (required by Sentence Transformers and Transformers)
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1

# NumPy with compatibility
numpy>=1.20.0,<2.0.0

# Request handling and retries
requests==2.31.0
tenacity==8.2.2

# Utilities
tqdm==4.65.0
psutil==5.9.5

# Optional development tools
memory-profiler==0.61.0